+++ dx7core/ringbuffer.cc
@@ -27,16 +27,16 @@ RingBuffer::RingBuffer() {
   wr_ix_ = 0;
 }
 
-int RingBuffer::BytesAvailable() {
+size_t RingBuffer::BytesAvailable() {
   return (wr_ix_ - rd_ix_) & (kBufSize - 1);
 }
 
-int RingBuffer::WriteBytesAvailable() {
+size_t RingBuffer::WriteBytesAvailable() {
   return (rd_ix_ - wr_ix_ - 1) & (kBufSize - 1);
 }
 
-int RingBuffer::Read(int size, uint8_t *bytes) {
-  int rd_ix = rd_ix_;
+size_t RingBuffer::Read(size_t size, uint8_t *bytes) {
+  size_t rd_ix = rd_ix_;
   SynthMemoryBarrier();  // read barrier, make sure data is committed before ix
   const size_t fragment_size = std::min(size, kBufSize - rd_ix);
   memcpy(bytes, buf_ + rd_ix, fragment_size);
@@ -48,12 +48,12 @@ int RingBuffer::Read(int size, uint8_t *bytes) {
   return size;
 }
 
-void RingBuffer::Write(const uint8_t *bytes, int size) {
-  unsigned int remaining = (unsigned int)size;
+void RingBuffer::Write(const uint8_t *bytes, size_t size) {
+  size_t remaining = size;
   while (remaining > 0) {
-    int rd_ix = rd_ix_;
-    int wr_ix = wr_ix_;
-    unsigned int space_available = (rd_ix - wr_ix - 1) & (kBufSize - 1);
+    size_t rd_ix = rd_ix_;
+    size_t wr_ix = wr_ix_;
+    size_t space_available = (rd_ix - wr_ix - 1) & (kBufSize - 1);
     if (space_available == 0) {
       struct timespec sleepTime;
       sleepTime.tv_sec = 0;
